[change] logic of CortinaTransitionMode
All code for handling CortinaTransitionMode is now in two places:
- playerPositionChanged
- fadeNow
The code for the transition has been adjusted to be the same in both
positions.
Fading in has been simlified to only look at the current track.
Fading out has been added and work on outro markes
Open items are:
- Current version does not cue to he correct start position
- Current version does not apply fade-in to the first track
- Use of transition time for fading when no intro/outro not implemented